Added code defining and	generating rows	and successors.
Originally, the	plan was to build a lookup table of successors for each
row, and use that to very efficiently generate increasingly small
successor sets for the partial grid as rows are added, leveraging sorted
vecs to compute set intersections in O(n) time.

However, it is now clear that there are	far too	many successors for each
row to do so within reasonable memory constraints, so it may be
necessary to generate successor sets after adding several rows, so that
the successor sets are sufficiently small, or it may be more efficient
to simply generate all rows by brute force on the fly, rather than store
anything more than the current partial grid in memory.

Signed-off-by: Oliver Calder <oliver@calder.dev>